restructured specific profile for consistency
