+++ src/js/index.mjs
@@ -4,12 +4,11 @@ import { registerUser } from "./auth/registration.mjs";
 import { loginUser } from "./auth/login.mjs";
 import { setup } from "./posts/setup.mjs";
 import { setupProfile } from "./profile/setup.mjs";
-import { loadProfile } from "./profile/setup.mjs";
 import { signout } from "./auth/logout.mjs";
 import { filterTags } from "./filter/filterTags.mjs";
 import { sortPosts } from "./sorter/sortType.mjs";
 import { setupSpecificPost } from "./posts/setupSpecific.mjs";
-import { setupSpecificProfile } from "./profile/getProfiles.mjs";
+import { setupSpecificProfile } from "./profile/setup.mjs";
 
 const path = window.location.pathname;
 console.log(path);
@@ -31,7 +30,6 @@ if (path === "/index.html") {
   setupSpecificPost();
   signout();
 } else if (path === "/profile.html") {
-  loadProfile();
   setupProfile();
   signout();
 } else if (path === "/profile-specific.html") {
