+++ src/js/profile/getSpecificProfile.mjs
@@ -2,8 +2,8 @@ import { API_PROFILE_URL } from "../auth/constants.mjs";
 import { getMyProfile } from "../handlers/getProfile.mjs";
 import { renderProfile } from "./renderProfile.mjs";
 import { renderProfilePosts } from "../posts/renderProfilePosts.mjs";
-import * as storage from "../storage/index.mjs";
 import { getPost } from "../handlers/getPost.mjs";
+
 /**
  * // function that sends get request with authorization token to retrieve a specific post from API server
  * Utilizes URLSearchParams to get the ID from the queryString
@@ -13,7 +13,7 @@ import { getPost } from "../handlers/getPost.mjs";
  * displaySpecificPosts()
  * ```
  */
-export async function loadProfile() {
+export async function getSpecificProfile() {
   const queryString = document.location.search;
   const params = new URLSearchParams(queryString);
   const name = params.get("name");
@@ -22,16 +22,11 @@ export async function loadProfile() {
   renderProfile(profile);
 }
 
-export async function loadSpecificProfilePosts() {
+export async function getSpecificProfilePosts() {
   const queryString = document.location.search;
   const params = new URLSearchParams(queryString);
   const user = params.get("name");
-  const postURL = `https://api.noroff.dev/api/v1/social/profiles/${user}/posts?_author=true&_reactions=true&_comments=true`;
-  const profilePosts = await getPost(postURL);
+  const SPECIFIC_PROFILEPOST_URL = `${API_PROFILE_URL}${user}/posts?_author=true&_reactions=true&_comments=true`;
+  const profilePosts = await getPost(SPECIFIC_PROFILEPOST_URL);
   renderProfilePosts(profilePosts);
-}
-
-export function setupSpecificProfile() {
-  loadProfile();
-  loadSpecificProfilePosts();
 }
